feat: fix analyse and formats

+++ app/FilamentTenant/Resources/ProductResource/RelationManagers/OptionsRelationManager.php
@@ -4,16 +4,15 @@ declare(strict_types=1);
 
 namespace App\FilamentTenant\Resources\ProductResource\RelationManagers;
 
-use App\Features\ECommerce\ColorPallete;
-use DB;
+use Domain\Product\Models\ProductOption;
 use Domain\Product\Models\ProductOptionValue;
 use Domain\Product\Models\ProductVariant;
+use Exception;
 use Filament\Resources\Form;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Resources\Table;
 use Filament\Tables;
 use Illuminate\Database\Eloquent\Builder;
-use Exception;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Facades\Log;
 
@@ -32,25 +31,8 @@ class OptionsRelationManager extends RelationManager
                 ->maxLength(100)
                 ->lazy()
                 ->columnSpan(2)
-                // ->columnSpan(
-                //     fn (Closure $get) => $get('../../is_custom') ? 1 : 2
-                // )
                 ->required(),
-            // \Filament\Forms\Components\Select::make('icon_type')
-            //     ->default('text')
-            //     ->required()
-            //     ->options(fn () => tenancy()->tenant?->features()->active(ColorPallete::class) ? [
-            //         'text' => 'Text',
-            //         'color_palette' => 'Color Palette',
-            //     ] : [
-            //         'text' => 'Text',
-            //     ]),
 
-            // \Filament\Forms\Components\ColorPicker::make('icon_value')
-            //     ->label(trans('Icon Value (HEX)')),
-            // ->hidden(fn (Closure $get) => !($get('icon_type') === 'color_palette' && $get('../../is_custom'))),
-            // ->hidden(fn (Closure $get) => !$get('../../is_custom'))
-            // ->reactive(),
         ]);
     }
 
@@ -83,6 +65,14 @@ class OptionsRelationManager extends RelationManager
                     ->translateLabel()
                     ->action(function (ProductOptionValue $record, Tables\Actions\Action $action): void {
                         try {
+                            if (! $record->productOption instanceof ProductOption) {
+                                $action
+                                    ->failureNotificationTitle(trans('The option value is unlinked from an option.'))
+                                    ->success();
+
+                                return;
+                            }
+
                             ProductVariant::where('product_id', $record->productOption->product_id)
                                 ->where(function (Builder $query) use ($record) {
                                     $query->whereJsonContains('combination', [['option_value_id' => $record->id]]);
@@ -106,9 +96,13 @@ class OptionsRelationManager extends RelationManager
                     ->action(function (Collection $records): void {
                         foreach ($records as $record) {
                             try {
+                                if (! isset($record->productOption)) {
+                                    return;
+                                }
+
                                 ProductVariant::where('product_id', $record->productOption->product_id)
                                     ->where(function (Builder $query) use ($record) {
-                                        $query->whereJsonContains('combination', [['option_value_id' => $record->id]]);
+                                        $query->whereJsonContains('combination', [['option_value_id' => $record->id ?? 0]]);
                                     })->delete();
 
                                 $record->delete();
